requirements.txt
discord.py
requests
beautifulsoup4
selenium
webdriver-manager
python-dotenv 